requirements.txt
anthropic
pandas
python-dotenv
markdown